Split `AnimationFile` related structs to separate header file
